requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
marshmallow==3.20.1
pyjwt==2.8.0
werkzeug==2.3.7
gunicorn==21.2.0
pytest==7.4.2
black==23.9.1
flake8==6.1.0 